Final changes to HGCPassive
