+++ SimG4CMS/HGCalTestBeam/plugins/HGCPassive.cc
@@ -4,21 +4,83 @@
 // Description: Main analysis class for HGCal Validation of G4 Hits
 ///////////////////////////////////////////////////////////////////////////////
 
-#include "SimG4CMS/HGCalTestBeam/interface/HGCPassive.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
 
+// to retreive hits
+#include "SimDataFormats/CaloHit/interface/PassiveHit.h"
+
+#include "SimG4Core/Notification/interface/BeginOfEvent.h"
+#include "SimG4Core/Notification/interface/BeginOfRun.h"
+#include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Notification/interface/Observer.h"
+#include "SimG4Core/Watcher/interface/SimProducer.h"
+
+#include "G4LogicalVolumeStore.hh"
+#include "G4PhysicalVolumeStore.hh"
+#include "G4Step.hh"
+#include "G4TransportationManager.hh"
+#include "G4TouchableHistory.hh"
+#include "G4Track.hh"
 #include "CLHEP/Units/GlobalPhysicalConstants.h"
 #include "CLHEP/Units/GlobalSystemOfUnits.h"
-#include "G4TransportationManager.hh"
 #include "DD4hep/Filter.h"
 
+#include <array>
 #include <cmath>
 #include <iomanip>
 #include <iostream>
+#include <map>
 #include <memory>
+#include <string>
+
 #include <utility>
+#include <vector>
 
 //#define EDM_ML_DEBUG
+
+class HGCPassive : public SimProducer,
+                   public Observer<const BeginOfRun *>,
+                   public Observer<const BeginOfEvent *>,
+                   public Observer<const G4Step *> {
+public:
+  HGCPassive(const edm::ParameterSet &p);
+  HGCPassive(const HGCPassive &) = delete;  // stop default
+  const HGCPassive &operator=(const HGCPassive &) = delete;
+  ~HGCPassive() override;
+
+  void produce(edm::Event &, const edm::EventSetup &) override;
+
+private:
+  // observer classes
+  void update(const BeginOfRun *run) override;
+  void update(const BeginOfEvent *evt) override;
+  void update(const G4Step *step) override;
+
+  // void endOfEvent(edm::PassiveHitContainer &HGCEEAbsE);
+  void endOfEvent(edm::PassiveHitContainer &hgcPH, unsigned int k);
+
+  typedef std::map<G4LogicalVolume *, std::pair<unsigned int, std::string>>::iterator volumeIterator;
+  G4VPhysicalVolume *getTopPV();
+  volumeIterator findLV(G4LogicalVolume *plv);
+  void storeInfo(
+      const volumeIterator itr, G4LogicalVolume *plv, unsigned int copy, double time, double energy, bool flag);
+
+private:
+  std::vector<std::string> LVNames_;
+  G4VPhysicalVolume *topPV_;
+  G4LogicalVolume *topLV_;
+  std::map<G4LogicalVolume *, std::pair<unsigned int, std::string>> mapLV_;
+  std::string motherName_;
+  int addlevel_;
+
+  // some private members for ananlysis
+  unsigned int count_;
+  bool init_;
+  std::map<std::pair<G4LogicalVolume *, unsigned int>, std::array<double, 3>> store_;
+};
 
 HGCPassive::HGCPassive(const edm::ParameterSet& p) : topPV_(nullptr), topLV_(nullptr), count_(0), init_(false) {
   edm::ParameterSet m_Passive = p.getParameter<edm::ParameterSet>("HGCPassive");
@@ -220,3 +282,8 @@ void HGCPassive::storeInfo(const HGCPassive::volumeIterator it,
                              << (itr->second)[2];
 #endif
 }
+
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+DEFINE_SIMWATCHER(HGCPassive);
