--- SimG4CMS/HGCalTestBeam/plugins/module.cc
@@ -1,5 +1,0 @@
-#include "SimG4CMS/HGCalTestBeam/interface/HGCPassive.h"
-#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-
-DEFINE_SIMWATCHER(HGCPassive);
